+++ CiFo-Sln/CIFO.Core/Infraestructure/ICloudStorageProvider.cs
@@ -8,7 +8,7 @@ namespace CIFO.Core.Infraestructure
 {
     public interface ICloudStorageProvider
     {
-        Task<string> Upload(int userId, Stream file, string contentType, string fileName);
+        Task<string> Upload(Int64 userId, Stream file, string contentType, string fileName);
 
     }
 }
