+++ CiFo-Sln/CiFo.Model/GovFolder/DocumentGcDto.cs
@@ -2,7 +2,7 @@
 {
     public class DocumentGcDto
     {
-        public int IdCitizen { get; set; }
+        public Int64 IdCitizen { get; set; }
         public string? UrlDocument { get; set; }
         public string? DocumentTitle { get; set; }
